requirements.txt
asgiref==3.8.1
Django==5.2.1
mysqlclient==2.2.4
pillow==10.4.0
sqlparse==0.5.0
typing_extensions==4.12.2
gunicorn==23.0.0
## The following requirements were added by pip freeze:
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
packaging==25.0
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0